Conflicts:
	application/modules/dashboard/controllers/media_stream.php
	application/modules/dashboard/views/index.php
	application/views/layout.php
	media/css/maybankdcms.css
	media/js/thekamarel.js

+++ application/modules/cms/controllers/cms.php
@@ -1,19 +1,52 @@
 <?php if ( ! defined('BASEPATH')) exit('No direct script access allowed');
 
 class Cms extends MY_Controller {
-
+     
+     public $user_role;
 	public function __construct()
 	{
 		parent::__construct();
+                $this->load->model('users_model');
 		$this->load->library(array('Shorturl', 'ciqrcode'));
 		$this->load->model(array('tag_model', 'product_model', 'campaign_model', 'shorturl_model', 'campaign_url_model'));
 		$this->load->helper('form');
 		$this->load->library('form_validation');
+                $this->user_role=$this->users_model->get_collection_detail(
+                                        array('role_collection_id'=>$this->session->userdata('role_id')));
 	}
 
     public function index()
     {
-    	$data['campaigns'] = $this->campaign_model->getAllArray();
+     if(IsRoleFriendlyNameExist($this->user_role,'Content Management_Campaign_View')||IsRoleFriendlyNameExist($this->user_role,'Content Management_Short_URL_View')
+        ||IsRoleFriendlyNameExist($this->user_role,'Content Management_Product_View')||IsRoleFriendlyNameExist($this->user_role,'Content Management_TAG_View'))
+     {
+     	$config['base_url'] = site_url('cms/index');
+		
+		$config['total_rows'] = $this->tag_model->count_record();
+		
+		$config['per_page'] = 10;
+		
+		$config["uri_segment"] = 3;
+	  
+		$config['next_link'] = 'Next';
+	  
+		$config['prev_link'] = 'Prev';
+	  
+		$config['first_link'] = 'First';
+	  
+		$config['last_link'] = 'Last';
+     
+		$config['cur_tag_open'] = '<b style="margin:0px 5px;">';
+	  
+		$config['cur_tag_close'] = '</b>';
+	  
+		$this->pagination->initialize($config);
+		
+		$offset = ($this->uri->segment(3)) ? $this->uri->segment(3) : 0;
+		
+    	$data['campaigns'] = $this->campaign_model->getAllArray($config['per_page'], $offset);
+    	
+    	$data['pagination'] = $this->pagination->create_links();
     	
     	$data['products'] = $this->product_model->get();
     	
@@ -24,11 +57,44 @@ class Cms extends MY_Controller {
         $data['cms_view'] = 'campaign_table';
         
         $this->load->view('cms/index',$data);
+     }
+     else
+     {
+          redirect('dashboard');
+     }
     }
      
     public function create_campaign()
     {
-    	$data['campaigns'] = $this->campaign_model->getAllArray();
+     if(IsRoleFriendlyNameExist($this->user_role,'Content Management_Campaign_View'))
+     {
+     	$config['base_url'] = site_url('cms/create_campaign');
+		
+		$config['total_rows'] = $this->tag_model->count_record();
+		
+		$config['per_page'] = 10;
+		
+		$config["uri_segment"] = 3;
+	  
+		$config['next_link'] = 'Next';
+	  
+		$config['prev_link'] = 'Prev';
+	  
+		$config['first_link'] = 'First';
+	  
+		$config['last_link'] = 'Last';
+     
+		$config['cur_tag_open'] = '<b style="margin:0px 5px;">';
+	  
+		$config['cur_tag_close'] = '</b>';
+	  
+		$this->pagination->initialize($config);
+		
+		$offset = ($this->uri->segment(3)) ? $this->uri->segment(3) : 0;
+     	
+    	$data['campaigns'] = $this->campaign_model->getAllArray($config['per_page'], $offset);
+    	
+    	$data['pagination'] = $this->pagination->create_links();
     	
     	$data['products'] = $this->product_model->get();
     	
@@ -83,17 +149,49 @@ class Cms extends MY_Controller {
         }
         
         $this->load->view('cms/index',$data);
+     }
+     else
+     {
+          redirect('cms');
+     }
     }
      
     public function create_tag(){
-    	
+     if(IsRoleFriendlyNameExist($this->user_role,'Content Management_TAG_View'))
+     {
     	$data['campaigns'] = '';
     	
     	$data['products'] = '';
     	
-    	$data['tags'] = $this->tag_model->get();
-    	
     	$data['urls'] = '';
+    	
+    	$config['base_url'] = site_url('cms/create_tag');
+		
+		$config['total_rows'] = $this->tag_model->count_record();
+		
+		$config['per_page'] = 10;
+		
+		$config["uri_segment"] = 3;
+	  
+		$config['next_link'] = 'Next';
+	  
+		$config['prev_link'] = 'Prev';
+	  
+		$config['first_link'] = 'First';
+	  
+		$config['last_link'] = 'Last';
+     
+		$config['cur_tag_open'] = '<b style="margin:0px 5px;">';
+	  
+		$config['cur_tag_close'] = '</b>';
+	  
+		$this->pagination->initialize($config);
+		
+		$offset = ($this->uri->segment(3)) ? $this->uri->segment(3) : 0;
+    	
+    	$data['tags'] = $this->tag_model->get($config['per_page'], $offset);
+    	
+    	$data['pagination'] = $this->pagination->create_links();
     	
     	$action = $this->input->get('action');
     	
@@ -101,6 +199,7 @@ class Cms extends MY_Controller {
         
         if ($this->input->server('REQUEST_METHOD') === "POST")
         {
+          if(IsRoleFriendlyNameExist($this->user_role,'Content Management_TAG_Create')){
 	        $arr = array();
 	        $arr['tag_name'] = $this->input->post('tag_name');
 	        $arr['user_id'] = $this->session->userdata('user_id');
@@ -117,10 +216,16 @@ class Cms extends MY_Controller {
 		        $this->session->set_flashdata('message_body', 'Please insert Tag name');
 	        }
 	        redirect('cms/create_tag');
+          }
+          else
+          {
+               redirect('cms/create_tag');
+          }
         }
         
         if ($action == 'delete')
         {
+          if(IsRoleFriendlyNameExist($this->user_role,'Content Management_TAG_Delete')){
 	        $id = $this->input->get('id');
 	        
 	        if ($id)
@@ -129,20 +234,98 @@ class Cms extends MY_Controller {
 	        }
 	        
 	        redirect('cms/create_tag');
+          }
+          else
+          {
+               redirect('cms/create_tag');
+          }
         }
         
         $this->load->view('cms/index',$data);
+     }
+     else
+     {
+          redirect('cms');
+     }
     }
      
     public function create_short_url()
     {
+     if(IsRoleFriendlyNameExist($this->user_role,'Content Management_Short_URL_View'))
+     {
     	$data['campaigns'] = $this->campaign_model->get();
     	
     	$data['products'] = '';
     	
     	$data['tags'] = '';
     	
-    	$data['urls'] = $this->campaign_url_model->get();
+    	
+    	// First Tab -----------------------------------------
+    	$cfg['base_url'] = site_url('cms/create_short_url');
+		
+		$cfg['total_rows'] = $this->campaign_url_model->count_record();
+		
+		$cfg['per_page'] = 10;
+		
+		$cfg["uri_segment"] = 3;
+	  
+		$cfg['next_link'] = 'Next';
+	  
+		$cfg['prev_link'] = 'Prev';
+	  
+		$cfg['first_link'] = 'First';
+	  
+		$cfg['last_link'] = 'Last';
+     
+		$cfg['cur_tag_open'] = '<b style="margin:0px 5px;">';
+	  
+		$cfg['cur_tag_close'] = '</b>';
+		
+		$cfg['suffix'] = '/firstTab';
+		
+		$cfg['first_url'] = site_url('cms/create_short_url/0/firstTab');
+	  
+		$this->pagination->initialize($cfg);
+		
+		$offset = ($this->uri->segment(3)) ? $this->uri->segment(3) : 0;
+
+    	$data['urls'] = $this->campaign_url_model->get($cfg['per_page'], $offset);
+    		
+    	$data['links'] = $this->pagination->create_links();
+    	
+    	// End of First Tab ------------------------------------
+    	
+    	// Second Tab ------------------------------------------
+    	
+    	$config['base_url'] = site_url('cms/create_short_url');
+		
+		$config['total_rows'] = $this->shorturl_model->count_record();
+		
+		$config['per_page'] = 10;
+		
+		$config["uri_segment"] = 3;
+	  
+		$config['next_link'] = 'Next';
+	  
+		$config['prev_link'] = 'Prev';
+	  
+		$config['first_link'] = 'First';
+	  
+		$config['last_link'] = 'Last';
+     
+		$config['cur_tag_open'] = '<b style="margin:0px 5px;">';
+	  
+		$config['cur_tag_close'] = '</b>';
+		
+		$config['suffix'] = '/secondTab';
+		
+		$config['first_url'] = site_url('cms/create_short_url/0/secondTab');
+	  
+		$this->pagination->initialize($config);
+		
+		$offset = ($this->uri->segment(3)) ? $this->uri->segment(3) : 0;
+    	
+    	$data['shorturls'] = $this->shorturl_model->get($config['per_page'], $offset);
     	
     	$action = $this->input->get('action');
     
@@ -206,18 +389,52 @@ class Cms extends MY_Controller {
 		
 		$data['code'] = substr( md5( time().uniqid().rand() ), 0, 6 );
         $data['cms_view'] = 'create_short_url';
+        $data['pagination'] = $this->pagination->create_links();
         $this->load->view('cms/index',$data);
+     }
+     else
+     {
+          redirect('cms');
+     }
     }
      
     public function create_product()
     {
+     if(IsRoleFriendlyNameExist($this->user_role,'Content Management_Product_View'))
+     {
     	$data['campaigns'] = '';
     	
     	$data['tags'] = '';
     	
-    	$data['products'] = $this->product_model->get();
-    	
     	$data['urls'] = '';
+    	
+    	$config['base_url'] = site_url('cms/create_product');
+		
+		$config['total_rows'] = $this->product_model->count_record();
+		
+		$config['per_page'] = 10;
+		
+		$config["uri_segment"] = 3;
+	  
+		$config['next_link'] = 'Next';
+	  
+		$config['prev_link'] = 'Prev';
+	  
+		$config['first_link'] = 'First';
+	  
+		$config['last_link'] = 'Last';
+     
+		$config['cur_tag_open'] = '<b style="margin:0px 5px;">';
+	  
+		$config['cur_tag_close'] = '</b>';
+	  
+		$this->pagination->initialize($config);
+		
+		$offset = ($this->uri->segment(3)) ? $this->uri->segment(3) : 0;
+    	
+    	$data['products'] = $this->product_model->get($config['per_page'], $offset);
+    	
+    	$data['pagination'] = $this->pagination->create_links();
     	
     	$action = $this->input->get('action');
         
@@ -225,6 +442,8 @@ class Cms extends MY_Controller {
         
         if ($this->input->server('REQUEST_METHOD') === "POST")
         {
+          if(IsRoleFriendlyNameExist($this->user_role,'Content Management_Product_Create'))
+          {
         	$params = array();
 	        $params = $this->input->post('product');
 	        $params['user_id'] = $this->session->userdata('user_id');
@@ -241,21 +460,37 @@ class Cms extends MY_Controller {
 		        $this->session->set_flashdata('message_body', 'Please insert Tag name');
 	        }
 	        redirect('cms/create_product');
+          }
+          else
+          {
+               redirect('cms');
+          }
         }
-        
-        if ($action == 'delete')
-        {
-	        $id = $this->input->get('id');
-	        
-	        if ($id)
-	        {
-		       $this->product_model->delete($id);
-	        }
-	        
-	        redirect('cms/create_product');
-        }
-        
-        $this->load->view('cms/index',$data);
+               if ($action == 'delete')
+               {
+                    if(IsRoleFriendlyNameExist($this->user_role,'Content Management_Product_Delete'))
+                    {
+                       $id = $this->input->get('id');
+                       
+                       if ($id)
+                       {
+                              $this->product_model->delete($id);
+                       }
+                       
+                       redirect('cms/create_product');
+                    }
+                    else
+                    {
+                        redirect('cms/create_product');
+                    }
+               }
+               
+          $this->load->view('cms/index',$data);
+     }
+     else
+     {
+          redirect('cms');
+     }
     }
     
     public function url()
